EXR texture support via OpenEXR, RGBA16Float material textures
- Add OpenEXR as Conan dependency for full EXR compression support
  (DWAA, DWAB, PXR24, B44, ZIP, PIZ, RLE)
- Unified load_image_float(): EXR via OpenEXR, HDR via stbi_loadf,
  LDR via stbi_load (no double-linearization from stbi_loadf gamma)
- Custom MemoryIStream for reading EXR from USD asset buffers
- Upgrade material texture array from RGBA8Unorm to RGBA16Float
- Extract float_to_half() into shared halfFloat.h
- Cache failed texture loads to avoid repeated decode attempts
- Add treasure chest model with EXR textures (metal, roughness, normal)

+++ core/include/core/rendering/preparedSceneData.h
@@ -36,7 +36,7 @@ struct PreparedSceneData {
 
     // Textures — non-owning pixel pointers, stable during frame
     struct TextureLayer {
-        const uint8_t* pixels;
+        const uint16_t* pixels;  // RGBA16Float (half-precision)
         uint32_t width;
         uint32_t height;
     };
